+++ src/libs/elements/Picture/types.d.ts
@@ -1,12 +1,15 @@
 import { GlobProps } from "../../../typings";
+import { WidthProps, HeightProps } from "styled-system";
 
 type srcSetProps = {
   media: string;
   src: string;
 };
 
-export interface PictureProps extends GlobProps {
+export interface PictureProps extends GlobProps, HTMLPictureElement {
   dataSrc?: string;
   alt?: string;
   srcSets?: srcSetProps[];
+  width?: WidthProps;
+  height?: HeightProps;
 }